Split code into multiple files, add tests for source json files, correct one json file
