+++ lorem/mikhail-gorbachev.json
@@ -6,9 +6,7 @@
     "As a young man, I really took to heart the Communist ideals. A young soul certainly cannot reject things like justice and equality. These were the goals proclaimed by the Communists. But in reality that terrible Communist experiment brought about repression of human dignity. Violence was used in order to impose that model on society. In the name of Communism we abandoned basic human values. So when I came to power in Russia I started to restore those values; values of \"openness\" and freedom.",
     "I believe in the cosmos. All of us are linked to the cosmos. Look at the sun. If there is no sun, then we cannot exist. So nature is my god. To me, nature is sacred. Trees are my temples and forests are my cathedrals.",
     "Preparing for my address I found in an old Russian encyclopedia a definition of \"peace\" as a \"commune\" — the traditional cell of Russian peasant life. I saw in that definition the people's profound understanding of peace as harmony, concord, mutual help, and cooperation. This understanding is embodied in the canons of world religions and in the works of philosophers from antiquity to our time.",
-    "I began my book about perestroika and the new thinking with the following words: \"We want to be understood\". After a while I felt that it was already happening. But now I would like once again to repeat those words here, from this world rostrum. Because to understand us really — to understand so as to believe us — proved to be not at all easy, owing to the immensity of the changes under way in our country. Their magnitude and character are such as to require in-depth analysis. Applying conventional wisdom to perestroika is unproductive. It is also futile and dangerous to set conditions, to say: We'll understand and believe you, as soon as you, the Soviet Union, come completely to resemble "
-    us
-    ", the West. No one is in a position to describe in detail what perestroika will finally produce. But it would certainly be a self-delusion to expect that perestroika will produce \"a copy\" of anything.",
+    "I began my book about perestroika and the new thinking with the following words: \"We want to be understood\". After a while I felt that it was already happening. But now I would like once again to repeat those words here, from this world rostrum. Because to understand us really — to understand so as to believe us — proved to be not at all easy, owing to the immensity of the changes under way in our country. Their magnitude and character are such as to require in-depth analysis. Applying conventional wisdom to perestroika is unproductive. It is also futile and dangerous to set conditions, to say: We'll understand and believe you, as soon as you, the Soviet Union, come completely to resemble \"us\", the West. No one is in a position to describe in detail what perestroika will finally produce. But it would certainly be a self-delusion to expect that perestroika will produce \"a copy\" of anything.",
     "A period of transition to a new quality in all spheres of society's life is accompanied by painful phenomena. When we were initiating perestroika we failed to properly assess and foresee everything. Our society turned out to be hard to move off the ground, not ready for major changes which affect people's vital interests and make them leave behind everything to which they had become accustomed over many years. In the beginning we imprudently generated great expectations, without taking into account the fact that it takes time for people to realize that all have to live and work differently, to stop expecting that new life would be given from above.",
     "I am an optimist and I believe that together we shall be able now to make the right historical choice so as not to miss the great chance at the turn of centuries and millenia and make the current extremely difficult transition to a peaceful world order. A balance of interests rather than a balance of power, a search for compromise and concord rather than a search for advantages at other people's expense, and respect for equality rather than claims to leadership — such are the elements which can provide the groundwork for world progress and which should be readily acceptable for reasonable people informed by the experience of the twentieth century. The future prospect of truly peaceful global politics lies in the creation through joint efforts of a single international democratic space in which States shall be guided by the priority of human rights and welfare for their own citizens and the promotion of the same rights and similar welfare elsewhere. This is an imperative of the growing integrity of the modern world and of the interdependence of its components.",
     "Have we not been able to cross the threshold of mistrust, though mistrust has not completely disappeared? Has not the political thinking in the world changed substantially? Does not most of the world community already regard weapons of mass destruction as unacceptable for achieving political objectives?",
